refactor(pymake): refactor pymake

+++ autotest/t002_test.py
@@ -72,11 +72,19 @@ def compile_code():
 def clean_up():
     # clean up downloaded directory
     print('Removing folder ' + swtpth)
-    shutil.rmtree(swtpth)
+    if os.path.isdir(swtpth):
+        shutil.rmtree(swtpth)
+
+    ext = ''
+    if sys.platform == 'win32':
+        ext = '.exe'
 
     # clean up target
     print('Removing ' + target)
-    os.remove(epth)
+    fpth = epth + ext
+    if os.path.isfile(fpth):
+        os.remove(fpth)
+
     return
 
 
@@ -98,27 +106,37 @@ def run_seawat(namepth, dst):
                                             model_ws=testpth, silent=True)
         if success:
             pymake.teardown(testpth)
+        else:
+            errmsg = 'could not run...{}'.format(os.path.basename(namepth))
     else:
         success = False
+        errmsg = '{} does not exist'.format(epth)
 
-
-    assert success is True
+    assert success, errmsg
 
     return
 
 
 def build_seawat_dependency_graphs():
-    # build dependencies output directory
-    if not os.path.exists(deppth):
-        os.makedirs(deppth)
+    if os.path.exists(epth):
 
-    # build dependency graphs
-    print('building dependency graphs')
-    pymake.visualize.make_plots(srcpth, deppth)
+        # build dependencies output directory
+        if not os.path.exists(deppth):
+            os.makedirs(deppth)
 
-    # test that the dependency figure for the SEAWAT main exists
-    findf = os.path.join(deppth, 'swt_v4.f.png')
-    assert os.path.isfile(findf) is True
+        # build dependency graphs
+        print('building dependency graphs')
+        pymake.visualize.make_plots(srcpth, deppth)
+
+        # test that the dependency figure for the SEAWAT main exists
+        findf = os.path.join(deppth, 'swt_v4.f.png')
+        success = os.path.isfile(findf)
+        assert success, 'could not find {}'.format(findf)
+    else:
+        success = False
+
+    assert success, 'could not build dependency graphs'
+
     return
